requirements.txt
pandas>=1.2.0
openpyxl~=3.1.0
pillow~=8.2.0
python-dotenv~=0.17.1
